[stdlib] Introduce Int128 and UInt128.
The type Int128 and UInt128 are integer types which are 128 bits wide.
The are just a 128 bits version of the existing integer types in the
stdlib.

The only operation which seems to not be supported by the 128 bits
integers is `dividingFullWidth` which needs 256 bits integers which
doesn't seem to be supported.

Includes modifications into Hasher to understand the new wider types.
To implement the hashing itself, each of the words of the types are
iterated and combined with the already existing methods.

Includes modifications to the integer tests to check that the 128 bits
integers are correctly handled in the same cases as the existing
types.