[WIP] PathSanitizingFileCheck will try to escape paths for Windows.
This is a rewrite of PathSanitizingFileCheck. At the moment it should be
not changed at all for Unix platforms or Windows platforms, but it
offers the opportunity to start adding `PATH((...))` markers in CHECK
lines which will be correctly escaped for the running platform.

The benefit of this change is the following:
- Unix will not accept Windows path separators by mistake (the current
  solution is not platform dependent).
- Using `PATH((...))` is slightly better than sprinkling `{{/|\\}}` or
  other variants in every path separator in the CHECK string.
- It will automatically support double escaped Windows separators as
  needed (the tests should already be marked with
  `--enable-yaml-compatibility`, and `PATH((...))` will do the right
  thing there).

The cons:
- It is not completely transparent for the Unix folks. You have to write
  `PATH((...))` explicitely. While heuristics to detect paths in the
  CHECK lines should be possible, I wanted a simpler solution first.